src/main/kotlin/com/dls/insuranceadvisorservice/domain/insurancelines/rules/LowIncomeRule.kt
package com.dls.insuranceadvisorservice.domain.insurancelines.rules

import com.dls.insuranceadvisorservice.domain.insurancelines.RiskScoreRule
import com.dls.insuranceadvisorservice.domain.riskprofile.RiskProfileForInsuranceLine
import com.dls.insuranceadvisorservice.domain.customer.Customer

class LowIncomeRule: RiskScoreRule {

    val LOW_INCOME = 25000;

     override fun execute(customer: Customer, riskProfileForInsuranceLine: RiskProfileForInsuranceLine){
        if(customer.income < LOW_INCOME){
            riskProfileForInsuranceLine.makeIneligible()
        }
    }
}