src/test/kotlin/com/dls/insuranceadvisorservice/domain/insurancelines/rules/LowIncomeRuleTest.kt
package com.dls.insuranceadvisorservice.domain.insurancelines.rules;

import com.dls.insuranceadvisorservice.domain.riskprofile.RiskProfileForInsuranceLine
import com.dls.insuranceadvisorservice.domain.customer.Customer
import org.junit.jupiter.api.Test


internal class LowIncomeRuleTest {

    private val rule =LowIncomeRule()

    @Test
    fun `Given a user with income lower than $25k, this rule must become the line of insurance INELIGIBLE insurance`() {
        //GIVEN
        val actualScore = 2
        val userProfile = givenUserProfile(2000)
        val riskProfileBaseLine = givenARiskProfileBaseLine(actualScore)
        //WHEN
        rule.execute(userProfile,riskProfileBaseLine);
        //THEN
        assert(riskProfileBaseLine.score == actualScore)
        assert(riskProfileBaseLine.insurancePlanStatus == RiskProfileForInsuranceLine.InsurancePlanStatus.INELIGIBLE)
    }

    @Test
    fun `Given a user with income above $25k, this rule mustn't deduct any point to the risk score of a line insurance`() {
        //GIVEN
        val actualScore = 2
        val userProfile = givenUserProfile(100001)
        val riskProfileBaseLine = givenARiskProfileBaseLine(actualScore)
        //WHEN
        rule.execute(userProfile,riskProfileBaseLine);
        //THEN
        assert(riskProfileBaseLine.score == actualScore)
        assert(riskProfileBaseLine.insurancePlanStatus == null)

    }

    private fun givenARiskProfileBaseLine(actualScore: Int) =
        RiskProfileForInsuranceLine(
            name= RiskProfileForInsuranceLine.Name.AUTO,
            score=actualScore)

    private fun givenUserProfile(income: Int) =
        Customer(
            age=30,
            dependents=1,
            income=income,
            maritalStatus = Customer.MaritalStatus.married,
            house = null,
            questionScore = listOf(1,1,0),
            vehicle = null
        )


}